requirements.txt

dj-database-url==0.5.0
Django==3.0.4
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-hitcount==1.3.1
django-markdown-deux==1.0.5
django-pagedown==2.0.3
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
Pillow==7.0.0



